gtr,parser/gotest: Add test 27

+++ pkg/gtr/event.go
@@ -46,7 +46,7 @@ type Event struct {
 
 	// Benchmarks
 	Iterations  int64
-	NsPerOp     int64
+	NsPerOp     float64
 	BytesPerOp  int64
 	AllocsPerOp int64
 }
